# Custom dotted_pointer implementation