--- a/src/lib.rs
+++ b/src/lib.rs
@@ -27,7 +27,9 @@ pub use crate::builtins::functions::Function;
 pub use crate::builtins::testers::Test;
 pub use crate::context::Context;
 pub use crate::errors::{Error, ErrorKind, Result};
-// Template and get_json_pointer are meant to be used internally only but is exported for test/bench.
+// Template, dotted_pointer and get_json_pointer are meant to be used internally only but is exported for test/bench.
+#[doc(hidden)]
+pub use crate::context::dotted_pointer;
 #[doc(hidden)]
 pub use crate::context::get_json_pointer;
 #[doc(hidden)]