… benches/json_pointer.rs
… src/renderer/processor.rs
… src/renderer/stack_frame.rs
… src/renderer/tests/square_brackets.rs