`get_json_pointer` converted a dotted pointer to an slashed pointer for serde jsons `pointer`.
I replaced the used of `pointer(get_json_pointer())` with an implementation of `dotted_pointer` what reduced a lot of allocations

as `get_json_pointer` was hidden, but a public function (for benchmarks), I marked it as deprecated.

benchmarks got replaced by a benchmark that uses `dotted_pointer` in a serde json value.

This improvement is not a breaking change.

Benchmarks before:
```
running 4 tests
test bench_big_loop_big_object                 ... bench:     184,662 ns/iter (+/- 2,075)
test bench_macro_big_object                    ... bench:     445,885 ns/iter (+/- 9,864)
test bench_macro_big_object_no_loop_macro_call ... bench:   1,052,077 ns/iter (+/- 152,796)
test bench_macro_big_object_no_loop_with_set   ... bench:   1,040,018 ns/iter (+/- 96,262)

test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured; 0 filtered out; finished in 10.07s

     Running benches/escaping.rs (target/release/deps/escaping-73cdee30956e38a2)

running 6 tests
test bench_escape_html_long         ... bench:         297 ns/iter (+/- 6)
test bench_escape_html_short        ... bench:         126 ns/iter (+/- 3)
test bench_escape_html_very_long    ... bench:       4,127 ns/iter (+/- 34)
test bench_escape_no_html_long      ... bench:         295 ns/iter (+/- 3)
test bench_escape_no_html_short     ... bench:          99 ns/iter (+/- 53)
test bench_escape_no_html_very_long ... bench:       2,374 ns/iter (+/- 48)

test result: ok. 0 passed; 0 failed; 0 ignored; 6 measured; 0 filtered out; finished in 11.67s

     Running benches/json_pointer.rs (target/release/deps/json_pointer-82582d912d61dd99)

running 2 tests
test bench_get_json_pointer          ... bench:         114 ns/iter (+/- 6)
test bench_get_json_pointer_with_map ... bench:         566 ns/iter (+/- 9)

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured; 0 filtered out; finished in 0.62s

     Running benches/templates.rs (target/release/deps/templates-2b7e84ae09e9002a)

running 2 tests
test big_table ... bench:   1,400,942 ns/iter (+/- 47,474)
test teams     ... bench:       4,410 ns/iter (+/- 47)

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured; 0 filtered out; finished in 3.20s

     Running benches/tera.rs (target/release/deps/tera-c384b6304625628d)

running 12 tests
test access_deep_object                        ... bench:       2,694 ns/iter (+/- 144)
test access_deep_object_with_literal           ... bench:       4,508 ns/iter (+/- 97)
test bench_build_inheritance_chains            ... bench:       2,181 ns/iter (+/- 147)
test bench_huge_loop                           ... bench:     593,874 ns/iter (+/- 20,272)
test bench_parsing_basic_template              ... bench:      44,001 ns/iter (+/- 423)
test bench_parsing_with_inheritance_and_macros ... bench:      69,565 ns/iter (+/- 534)
test bench_rendering_basic_template            ... bench:       3,626 ns/iter (+/- 112)
test bench_rendering_inheritance_and_macros    ... bench:       4,094 ns/iter (+/- 75)
test bench_rendering_only_inheritance          ... bench:         933 ns/iter (+/- 17)
test bench_rendering_only_macro_call           ... bench:       3,342 ns/iter (+/- 35)
test bench_rendering_only_parent               ... bench:         398 ns/iter (+/- 7)
test bench_rendering_only_variable             ... bench:         779 ns/iter (+/- 6)

test result: ok. 0 passed; 0 failed; 0 ignored; 12 measured; 0 filtered out; finished in 7.48s
```

Benchmarks after:
```
running 4 tests
test bench_big_loop_big_object                 ... bench:     186,405 ns/iter (+/- 1,487)
test bench_macro_big_object                    ... bench:     371,849 ns/iter (+/- 10,331)
test bench_macro_big_object_no_loop_macro_call ... bench:   1,021,208 ns/iter (+/- 95,960)
test bench_macro_big_object_no_loop_with_set   ... bench:   1,028,662 ns/iter (+/- 39,664)

test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured; 0 filtered out; finished in 6.73s

     Running benches/dotted_pointer.rs (target/release/deps/dotted_pointer-4b5732f0741af2c5)

running 2 tests
test bench_get_dotted_pointer          ... bench:         285 ns/iter (+/- 2)
test bench_get_dotted_pointer_with_map ... bench:         288 ns/iter (+/- 1)

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured; 0 filtered out; finished in 4.04s

     Running benches/escaping.rs (target/release/deps/escaping-73cdee30956e38a2)

running 6 tests
test bench_escape_html_long         ... bench:         284 ns/iter (+/- 11)
test bench_escape_html_short        ... bench:         116 ns/iter (+/- 6)
test bench_escape_html_very_long    ... bench:       4,128 ns/iter (+/- 56)
test bench_escape_no_html_long      ... bench:         454 ns/iter (+/- 15)
test bench_escape_no_html_short     ... bench:         139 ns/iter (+/- 2)
test bench_escape_no_html_very_long ... bench:       1,400 ns/iter (+/- 8)

test result: ok. 0 passed; 0 failed; 0 ignored; 6 measured; 0 filtered out; finished in 3.98s

     Running benches/templates.rs (target/release/deps/templates-2b7e84ae09e9002a)

running 2 tests
test big_table ... bench:   1,432,697 ns/iter (+/- 76,929)
test teams     ... bench:       3,578 ns/iter (+/- 119)

test result: ok. 0 passed; 0 failed; 0 ignored; 2 measured; 0 filtered out; finished in 4.55s

     Running benches/tera.rs (target/release/deps/tera-c384b6304625628d)

running 12 tests
test access_deep_object                        ... bench:       2,071 ns/iter (+/- 59)
test access_deep_object_with_literal           ... bench:       3,770 ns/iter (+/- 33)
test bench_build_inheritance_chains            ... bench:       2,035 ns/iter (+/- 30)
test bench_huge_loop                           ... bench:     509,551 ns/iter (+/- 13,032)
test bench_parsing_basic_template              ... bench:      43,277 ns/iter (+/- 590)
test bench_parsing_with_inheritance_and_macros ... bench:      68,536 ns/iter (+/- 857)
test bench_rendering_basic_template            ... bench:       2,830 ns/iter (+/- 61)
test bench_rendering_inheritance_and_macros    ... bench:       3,554 ns/iter (+/- 57)
test bench_rendering_only_inheritance          ... bench:         914 ns/iter (+/- 19)
test bench_rendering_only_macro_call           ... bench:       2,839 ns/iter (+/- 85)
test bench_rendering_only_parent               ... bench:         384 ns/iter (+/- 14)
test bench_rendering_only_variable             ... bench:         595 ns/iter (+/- 18)

test result: ok. 0 passed; 0 failed; 0 ignored; 12 measured; 0 filtered out; finished in 6.58s
```