@Keats 2023-02-16
(FYI I don't know if you are familiar with it but https://github.com/BurntSushi/cargo-benchcmp is great to see the changes visually)